+++ js/button/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/button
 
+## 0.5.9
+
+### Dependency updates
+
+- @gradio/upload@0.16.13
+- @gradio/client@1.16.0
+- @gradio/image@0.22.15
+
 ## 0.5.8
 
 ### Dependency updates