+++ js/component-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @self/component-test
 
+## 0.5.29
+
+### Dependency updates
+
+- @gradio/video@0.14.23
+
 ## 0.5.28
 
 ### Dependency updates