+++ js/model3D/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/model3d
 
+## 0.14.23
+
+### Dependency updates
+
+- @gradio/upload@0.16.13
+- @gradio/client@1.16.0
+
 ## 0.14.22
 
 ### Dependency updates