+++ js/multimodaltextbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/multimodaltextbox
 
+## 0.10.15
+
+### Dependency updates
+
+- @gradio/upload@0.16.13
+- @gradio/video@0.14.23
+- @gradio/client@1.16.0
+- @gradio/image@0.22.15
+
 ## 0.10.14
 
 ### Fixes