+++ js/video/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/video
 
+## 0.14.23
+
+### Dependency updates
+
+- @gradio/upload@0.16.13
+- @gradio/client@1.16.0
+- @gradio/image@0.22.15
+
 ## 0.14.22
 
 ### Dependency updates